removed earn context and aave context because we found them obsolete
