+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import { WithConnection } from 'components/connectWallet/ConnectWallet'
 import { DeferedContextProvider } from 'components/DeferedContextProvider'
 import { LandingPageLayout } from 'components/Layouts'
-import { earnContext, EarnContextProvider } from 'features/earn/EarnContextProvider'
+import { aaveContext, AaveContextProvider } from 'features/aave/AaveContextProvider'
 import { Survey } from 'features/survey'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import React from 'react'
@@ -16,14 +16,14 @@ export const getStaticProps = async ({ locale }: { locale: string }) => ({
 
 function LandingPage() {
   return (
-    <EarnContextProvider>
-      <DeferedContextProvider context={earnContext}>
+    <AaveContextProvider>
+      <DeferedContextProvider context={aaveContext}>
         <WithConnection>
           <HomepageView />
           <Survey for="homepage" />
         </WithConnection>
       </DeferedContextProvider>
-    </EarnContextProvider>
+    </AaveContextProvider>
   )
 }
